Let the leader handle the retry count of checkpoints
Fixes #1598. The worker was reducing the retry count because of some
stray state left around from the previous run. This commit moves that
check over to the leader, since the retry count should ideally be
state that only the leader knows or cares about.

This also fixes a bug where checkpointed jobs were retried too often.